requirements.txt
pyperclip
openai
python-dotenv